Final push

+++ remove_random.py
@@ -3,7 +3,7 @@ import random
 
 directory = "/local_datasets/imsi/CelebV/videos/"
 
-for i in range(1, 51):
+for i in range(1, 10001):
     file_name = f"{i:05}.mp4"
     file_path = os.path.join(directory, file_name)
     
@@ -13,15 +13,25 @@ for i in range(1, 51):
     else:
         print(f"File not found: {file_path}")
 
-all_files = [os.path.join(directory, f) for f in os.listdir(directory) if os.path.isfile(os.path.join(directory, f))]
-
-files_to_delete = random.sample(all_files, 30616)
-
-for file_path in files_to_delete:
-    try:
+for i in range(14000, 35667):
+    file_name = f"{i:05}.mp4"
+    file_path = os.path.join(directory, file_name)
+    
+    if os.path.exists(file_path):
         os.remove(file_path)
         print(f"Deleted: {file_path}")
-    except Exception as e:
-        print(f"Error deleting {file_path}: {e}")
+    else:
+        print(f"File not found: {file_path}")
 
-print("Deletion complete.")
+# all_files = [os.path.join(directory, f) for f in os.listdir(directory) if os.path.isfile(os.path.join(directory, f))]
+# 
+# files_to_delete = random.sample(all_files, 30616)
+# 
+# for file_path in files_to_delete:
+#     try:
+#         os.remove(file_path)
+#         print(f"Deleted: {file_path}")
+#     except Exception as e:
+#         print(f"Error deleting {file_path}: {e}")
+# 
+# print("Deletion complete.")
